Makefile

+++ Libft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: nvan-den <nvan-den@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/25 18:53:08 by nvan-den          #+#    #+#             */
-/*   Updated: 2022/10/26 15:55:56 by nvan-den         ###   ########.fr       */
+/*   Updated: 2022/10/27 12:20:06 by nvan-den         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 char	*ft_strnstr(const char *str, const char *to_find, size_t len)
 {
-	int		i;
-	int		j;
-	char	*d;
+	size_t		i;
+	size_t		j;
+	char		*d;
 
 	i = 0;
 	while (1)
